+++ src/graph.c
@@ -1,4 +1,5 @@
 #include <stdlib.h>
+#include <stdint.h>
 #include <stdio.h>
 #include <stdbool.h>
 
@@ -13,29 +14,29 @@
 // This edge struct acts as a linked list node
 typedef struct edge
 {
-    int destination;   /* destination vertex of edge a.k.a an adjascent vertex */
-    int weight;        /* edge weight */
-    struct edge *next; /* next edge in list... I have to use the struct keyboard because this is recursive */
+    uint64_t destination; /* destination vertex of edge a.k.a an adjascent vertex */
+    uint64_t weight;      /* edge weight */
+    struct edge *next;    /* next edge in list... I have to use the struct keyboard because this is recursive */
 } edge;
 
 // Our adjacency list is an array of linked lists called edges. The index of the edges array represents the id of the
 typedef struct graph
 {
-    edge *edges[MAXV + 1]; /* adjacency info */
-    int degree[MAXV + 1];  /* outdegree of each vertex */
-    int number_vertices;   /* number of vertices in graph */
-    int number_edges;      /* number of edges in graph */
-    bool is_directed;      /* Is the graph directed */
+    edge *edges[MAXV + 1];     /* adjacency info */
+    uint64_t degree[MAXV + 1]; /* outdegree of each vertex */
+    uint64_t number_vertices;  /* number of vertices in graph */
+    uint64_t number_edges;     /* number of edges in graph */
+    bool is_directed;          /* Is the graph directed */
 } graph;
 
 void initialize_graph(graph *g, bool directed);
 void read_graph(graph *g, bool directed);
-void insert_edge(graph *g, int x, int y, bool directed);
+void insert_edge(graph *g, uint64_t x, uint64_t y, bool directed);
 void print_graph(graph *g);
 
 void initialize_graph(graph *g, bool directed)
 {
-    for (int i = 1; i < MAXV; i++)
+    for (uint64_t i = 1; i < MAXV; i++)
     {
         g->edges[i] = NULL;
         g->degree[i] = 0;
@@ -47,22 +48,22 @@ void initialize_graph(graph *g, bool directed)
 
 void read_graph(graph *g, bool is_directed)
 {
-    int m;    /* number of edges */
-    int x, y; /* vertices in edge (x, y) */
+    uint64_t m;    /* number of edges */
+    uint64_t x, y; /* vertices in edge (x, y) */
 
     initialize_graph(g, is_directed);
 
-    scanf("%d %d", &(g->number_vertices), &m);
+    scanf("%lu %lu", &(g->number_vertices), &m);
 
-    for (int i = 1; i <= m; i++)
+    for (uint64_t i = 1; i <= m; i++)
     {
-        scanf("%d %d", &x, &y);
+        scanf("%lu %lu", &x, &y);
         insert_edge(g, x, y, is_directed);
     }
 }
 
 // Inserts an edge from source to destination in the adjascency list of graph g. If the edge is not directed, it adds source -> destination and destination -> but only increments th edge count once. Interestingly, weight is not yet implemented. This function would likely need to be modified to support weights on our edges
-void insert_edge(graph *g, int source, int destination, bool is_directed)
+void insert_edge(graph *g, uint64_t source, uint64_t destination, bool is_directed)
 {
     edge *new_edge;                  /* temporary pointer */
     new_edge = malloc(sizeof(edge)); /* allocate edge storage */
@@ -90,24 +91,15 @@ void print_graph(graph *g)
 {
     edge *p; /* temporary pointer */
 
-    for (int i = 1; i <= g->number_vertices; i++)
+    for (uint64_t i = 1; i <= g->number_vertices; i++)
     {
-        printf("%d: ", i);
+        printf("%lu: ", i);
         p = g->edges[i];
         while (p != NULL)
         {
-            printf(" %d", p->destination);
+            printf(" %lu", p->destination);
             p = p->next;
         }
         printf("\n");
     }
 }
-
-// int main()
-// {
-//     graph *g;
-//     g = malloc(sizeof(graph));
-//     read_graph(g, false);
-//     print_graph(g);
-//     return 0;
-// }
